test: Add golden files to enterprise cli
